Remove UI string resources from library.
These string resources were used to set the title and buttons of
NumberPickerDialog but that doesn't work well for localization as the
library can't provide these string resources in all languages that a
user of the library might want to localize an app to, and furthermore
would the proper localizaton also sometimes depend on context.

Instead, require these strings to be passed as arguments to the
NumberPickerDialog constructor. Even for non-localized apps this
provides greater flexibility for the user of the library.